+++ components/mojito/src/main/java/org/limewire/mojito/routing/Vendor.java
@@ -83,7 +83,7 @@ public class Vendor implements Serializable {
         
         int id = 0;
         for(char c : chars) {
-            id = (id << 8) | (int)(c & 0xFF);
+            id = (id << 8) | (c & 0xFF);
         }
         return id;
     }
